move getTopology to topology file
