+++ src/change_stream.ts
@@ -11,12 +11,11 @@ import {
   now,
   maybePromise,
   MongoDBNamespace,
-  Callback,
-  getTopology
+  Callback
 } from './utils';
 import type { ReadPreference } from './read_preference';
 import type { Timestamp, Document } from './bson';
-import type { Topology } from './sdam/topology';
+import { getTopology, Topology } from './sdam/topology';
 import type { OperationParent } from './operations/command';
 import type { CollationOptions } from './cmap/wire_protocol/write_command';
 const kResumeQueue = Symbol('resumeQueue');
