+++ src/operations/distinct.ts
@@ -1,9 +1,10 @@
 import { Aspect, defineAspects } from './operation';
 import { CommandOperation, CommandOperationOptions } from './command';
-import { decorateWithCollation, decorateWithReadConcern, Callback, getTopology } from '../utils';
+import { decorateWithCollation, decorateWithReadConcern, Callback } from '../utils';
 import type { Document } from '../bson';
 import type { Server } from '../sdam/server';
 import type { Collection } from '../collection';
+import { getTopology } from '../sdam/topology';
 
 /** @public */
 export type DistinctOptions = CommandOperationOptions;
